+++ src/CustomizeCritter/HappyEffects.cs
@@ -17,28 +17,28 @@ namespace CustomizeCritter
             {
                 ___happyWildEffect.SelfModifiers.Clear();
                 foreach (var attribute in CustomizeCritterState.StateManager.State.happyWildEffect)
-                    ___happyWildEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY.NAME, attribute.multiplier ?? false));
+                    ___happyWildEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY_WILD.NAME, attribute.multiplier ?? false));
             }
             
             if (CustomizeCritterState.StateManager.State.happyTameEffect != null)
             {
                 ___happyTameEffect.SelfModifiers.Clear();
                 foreach (var attribute in CustomizeCritterState.StateManager.State.happyTameEffect)
-                    ___happyTameEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY.NAME, attribute.multiplier ?? false));
+                    ___happyTameEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY_TAME.NAME, attribute.multiplier ?? false));
             }
             
             if (CustomizeCritterState.StateManager.State.unhappyWildEffect != null)
             {
                 ___unhappyWildEffect.SelfModifiers.Clear();
                 foreach (var attribute in CustomizeCritterState.StateManager.State.unhappyWildEffect)
-                    ___unhappyWildEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY.NAME, attribute.multiplier ?? false));
+                    ___unhappyWildEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY_WILD.NAME, attribute.multiplier ?? false));
             }
             
             if (CustomizeCritterState.StateManager.State.unhappyTameEffect != null)
             {
                 ___unhappyTameEffect.SelfModifiers.Clear();
                 foreach (var attribute in CustomizeCritterState.StateManager.State.unhappyTameEffect)
-                    ___unhappyTameEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY.NAME, attribute.multiplier ?? false));
+                    ___unhappyTameEffect.Add(new AttributeModifier(attribute.id, attribute.value, STRINGS.CREATURES.MODIFIERS.HAPPY_TAME.NAME, attribute.multiplier ?? false));
             }
         }
     }