+++ src/CustomizeElements/Changelog.md
@@ -1,9 +1,12 @@
 # Changelog
 
+## [1.0.4.0]
+- added support of sublimate settings
+
 ## [1.0.3.0] U45-545357-S
 - fixed crash with custom tags
 
-## [1.0.1.1] U41-498381-S U45-545357-S
+## [1.0.1.1] U41-498381-S
 - added access to modifiers
 
 ## [1.0.0.0] EX1 S14-471883-S