+++ src/CustomizePlants/Changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [1.0.42.0] 
+- small patch to increase compatiblity with Fast Track
+
 ## [1.0.41.0] U48-568201-S
 - fixed CheatFlowerVase, it no longer consumes material